--- .env
@@ -1,3 +1,0 @@
-NEXT_PUBLIC_SUPABASE_URL=https://btxflfkyfpbknsjnsmzw.supabase.co
-NEXT_PUBLIC_SUPABASE_ANON_KEY=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6ImJ0eGZsZmt5ZnBia25zam5zbXp3Iiwicm9sZSI6ImFub24iLCJpYXQiOjE3MTE2Nzc1MjgsImV4cCI6MjAyNzI1MzUyOH0.6bF7znGBj2Bn4i6sL63G9ms1RAF9dqJEG5M9cAj4FCQ
-NEXT_PUBLIC_KAKAO_MAP_API_KEY = 41c82cfaf90068132950d71292235e2b
