--- src/components/chat/test.ts
@@ -1,1 +1,0 @@
-// sdfsdf
