--- src/components/room/test.ts
@@ -1,1 +1,0 @@
-// sdfsdf
